Compile/link a generated syscall-veneer code, syscall-veneer.s
It seems that there is no reasonable way to use gas and nasm
together in the same project. Thus, the default assembly dialect
in the project is nasm and gas is compiled using a custom command.
This is ugly. :(

Currently, syscall-veneer.s is just a mock-up and should be replaced
to the real, generated syscall-veneer of an application.